Obstacles avoiding
The intention is to plan a path to avoid obstacles from starting point to the goal. The goal, the obstacles, the starting position of the path planning are already known.
If the expected goal point cannot be achieved, it may also be related to the improper setting of initial gain coefficient and po.
There may have obstacles near the target point, which leads to the problem of goal non reachable with obstacles nearby (GNRON). A new repulsion function is introduced. Based on the original repulsion field, it adds the influence of the distance between the target and the object. When the object is close to the target, the repulsion field will increase, but the distance will decrease, so it can drag the repulsion field.